add pre-processing code delete data and useless part

+++ data_utils.py
@@ -38,8 +38,8 @@ def get_samples_and_labels(settings):
         # always load eICU
         samples, pdf, labels = get_data('eICU_task', {})
         # del samples, labels
-        train, vali, test = samples['train'].astype(float), samples['vali'].astype(float), samples['test'].astype(float)
-        train_labels, vali_labels, test_labels = labels['train'].astype(float), labels['vali'].astype(float), labels['test'].astype(float)
+        train, vali, test = samples['train'], samples['vali'], samples['test']
+        train_labels, vali_labels, test_labels = labels['train'], labels['vali'], labels['test']
         assert train_labels.shape[1] == settings['cond_dim']
         # normalise to between -1, 1
         train, vali, test = normalise_data(train, vali, test)
@@ -288,8 +288,8 @@ def eICU_task(predict_label=False):
     path = './experiments/data/eicu.data.npy'
     data = np.load(path).item()
     # convert it into similar format
-    labels = {'train': data['Y_train'].astype(np.float), 'vali': data['Y_vali'].astype(np.float), 'test': data['Y_test'].astype(np.float)}
-    samples = {'train': data['X_train'].astype(np.float), 'vali': data['X_vali'].astype(np.float), 'test': data['X_test'].astype(np.float)}
+    labels = {'train': data['Y_train'], 'vali': data['Y_vali'], 'test': data['Y_test']}
+    samples = {'train': data['X_train'], 'vali': data['X_vali'], 'test': data['X_test']}
     # reshape
     #for (k, X) in samples.items():
     #    samples[k] = X.reshape(-1, 16, 4)
